refact(explorer): useIndividual hook
