+++ src/modules/manager/hooks.js
@@ -5,6 +5,7 @@ import { RESOURCE_EVERYTHING, viewDropBox } from "bento-auth-js";
 import { useHasResourcePermissionWrapper } from "@/hooks";
 import { useService } from "@/modules/services/hooks";
 import { fetchDropBoxTree } from "./actions";
+import { fetchIndividual } from "@/modules/metadata/actions";
 
 
 export const useDropBox = () => {
@@ -21,4 +22,22 @@ export const useDropBox = () => {
     }, [dispatch, dropBox, hasPermission]);
 
     return useSelector((state) => state.dropBox);
+};
+
+
+export const useIndividual = (individualID) => {
+    const dispatch = useDispatch();
+
+    const metadataService = useService("metadata");
+    const individuals = useSelector((state) => state.individuals.itemsByID);
+
+    useEffect(() => {
+        if (metadataService && individualID) {
+            // If we've loaded the metadata service, and we have an individual selected (or the individual ID changed),
+            // we should load individual data.
+            dispatch(fetchIndividual(individualID)).catch(console.error);
+        }
+    }, [dispatch, metadataService, individualID]);
+
+    return individuals[individualID];
 };
